+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ## v2.1.8
 * Added Stream (also StreamCodec) support for PersistedParser
+* Added flatten parameter for PersistedParser and ConfiguratorParser